2d - Texture2DArray
The texture class now creates a real array of textures, which can then be accessed by their index in the pixel shader.

The problem is, the textures are drawn only in their original size and cannot be made to fit their quads, in case the quads are smaller then the textures. Need to be fixed.