2d - Texture2DArray
Textures now can be loaded into the array from the file list.
Number of the textures to be loaded is defined by the qty of the files in the list.
Textures and animations are drawn normally.
There is problem, however, with the first instanced item not receiving its right animation frame. Look into it later.